fix: resolve review observations on corner validation, itoa safety, source labels, and seed fixture
- #1: firstStart now skips catalog_enumerated corners to avoid false
  positives in wrap-around overlap validation
- #2: itoa accepts uint to prevent empty string on negative input;
  cascaded uint through layoutFromGT7Info, layoutID, enumeratedCorners
- #3: regenerated gt7_official_seed_catalog.json fixture to reflect
  current seed (gt7_layout_XXXX IDs, enumerated corners, gt7Info sources)
- #4: officialNewsSource uses correct sourceType
  official_gran_turismo_news instead of delegating to
  officialTracklistAssetSource

+++ internal/tracks/catalog_test.go
@@ -337,7 +337,7 @@ func TestGT7OfficialSeedCatalogSourceOfTruthValidatesWithEnumeratedCorners(t *te
 	}
 	for i, corner := range layout.Corners {
 		want := i + 1
-		if corner.Number != want || corner.Name != "Corner "+itoa(want) || corner.DefinitionMode != CornerDefinitionCatalogEnumerated {
+		if corner.Number != want || corner.Name != "Corner "+itoa(uint(want)) || corner.DefinitionMode != CornerDefinitionCatalogEnumerated {
 			t.Fatalf("unexpected enumerated corner %d: %+v", want, corner)
 		}
 		if corner.StartMeters != 0 || corner.ApexMeters != 0 || corner.EndMeters != 0 {
